+++ Nexd/lib/openapi/NexdClient/Classes/OpenAPIs/Models/HelpList.swift
@@ -16,16 +16,16 @@ public struct HelpList: Codable {
         case canceled = "canceled"
         case completed = "completed"
     }
-    public var id: Int
+    public var id: Int64
     public var ownerId: String?
     public var owner: User?
     public var createdAt: Date
     public var updatedAt: Date
     public var status: Status? = .active
     public var helpRequests: [HelpRequest]
-    public var helpRequestsIds: [Int]
+    public var helpRequestsIds: [Int64]
 
-    public init(id: Int, ownerId: String?, owner: User?, createdAt: Date, updatedAt: Date, status: Status?, helpRequests: [HelpRequest], helpRequestsIds: [Int]) {
+    public init(id: Int64, ownerId: String?, owner: User?, createdAt: Date, updatedAt: Date, status: Status?, helpRequests: [HelpRequest], helpRequestsIds: [Int64]) {
         self.id = id
         self.ownerId = ownerId
         self.owner = owner
