+++ Nexd/lib/openapi/NexdClient/Classes/OpenAPIs/Models/HelpRequestArticle.swift
@@ -11,14 +11,14 @@ import Foundation
 public struct HelpRequestArticle: Codable { 
 
 
-    public var articleId: Int?
-    public var articleCount: Int?
-    public var id: Int?
+    public var articleId: Int64?
+    public var articleCount: Int64?
+    public var id: Int64?
     public var article: Article?
     public var articleDone: Bool?
     public var helpRequest: HelpRequest?
 
-    public init(articleId: Int?, articleCount: Int?, id: Int?, article: Article?, articleDone: Bool?, helpRequest: HelpRequest?) {
+    public init(articleId: Int64?, articleCount: Int64?, id: Int64?, article: Article?, articleDone: Bool?, helpRequest: HelpRequest?) {
         self.articleId = articleId
         self.articleCount = articleCount
         self.id = id
